Generalize types from `AsRef<str>` to `AsRef<OsStr>`
In Rust, `&str` represents a slice into a UTF-8 Unicode string.
`&OsStr` represents something which could be validly used as a pathname,
and which doesn't necessarily have any specific encoding.

If we're being obsessive about presenting an ideal public API, we
actually want the latter, not the former.  Most callers will not even
notice this change because we use `AsRef<OsStr>` to allow conversions
from other string representations.