remove NulError from error type
while the NulError is nice in that it includes where the NUL is, it
prevents us from having the same API on Windows as on Unix, and
doesn't really add enough value to overcome that, so it's now
replaced with a nullary Error variant 'NullByteInArgument'.

Bumping version to 0.4.0 because this is a breaking change.